Massive cleanup and fixes
