+++ .eslintignore
@@ -1,0 +1,3 @@
+/.next
+/node_modules
+/public
