--- css/form.css
@@ -1,39 +1,0 @@
-form {
-  width: 90%;
-  margin: auto;
-  max-width: 550px;
-}
-input,
-form button,
-label {
-  display: block;
-}
-form button,
-input,
-textarea {
-  outline: none;
-}
-input,
-textarea {
-  border: 1px solid rgb(199, 199, 199);
-  border-radius: 10px;
-  padding: 10px;
-  font-size: 90%;
-  width: 100%;
-  height: 30px;
-  color: rgb(53, 53, 53);
-}
-textarea {
-  height: 50px;
-}
-label {
-  margin-top: 10px;
-}
-form button {
-  --accent: rgb(0, 162, 255);
-  margin-top: 20px;
-}
-
-.form-container {
-  width: 90%;
-}
