--- jest.setup.js
@@ -1,6 +1,0 @@
-import { configure } from "enzyme";
-import Adapter from "enzyme-adapter-react-16";
-
-configure({
-  adapter: new Adapter(),
-});
