+++ pages/api/todos/[id].ts
@@ -1,0 +1,3 @@
+import type { NextApiRequest, NextApiResponse } from "next";
+
+export default (req: NextApiRequest, res: NextApiResponse) => {};
