--- vercel.json
@@ -1,10 +1,0 @@
-{
-  "env": {
-    "MONGODB_URI": "@next_example_mongodb_uri"
-  },
-  "build": {
-    "env": {
-      "MONGODB_URI": "@next_example_mongodb_uri"
-    }
-  }
-}
